+++ packages/now-static-build/src/frameworks.ts
@@ -1,22 +1,11 @@
-import {
-  readdir,
-  stat,
-  readFile,
-  unlink,
-  writeFile,
-  copyFile,
-  mkdir,
-} from 'fs';
+import { readdir, stat, readFile, unlink } from 'fs';
 import { promisify } from 'util';
 import { join } from 'path';
 import { Route } from '@now/build-utils';
-import { createGatsbyConfig } from './gatsby-generators';
+import { injectGatsbyConfig } from './utils/gatsby-config';
 
 const readirPromise = promisify(readdir);
 const readFilePromise = promisify(readFile);
-const writeFilePromise = promisify(writeFile);
-const mkdirPromise = promisify(mkdir);
-const copyFilePromise = promisify(copyFile);
 const statPromise = promisify(stat);
 const unlinkPromise = promisify(unlink);
 const isDir = async (file: string): Promise<boolean> =>
@@ -57,42 +46,7 @@ export const frameworks: Framework[] = [
       }
     },
     beforeBuildHook: async entrypointDir => {
-      let gatsbyConfigUser = false;
-      try {
-        await copyFilePromise(
-          join(entrypointDir, 'gatsby-config.js'),
-          join(entrypointDir, 'gatsby-config-user.js')
-        );
-        gatsbyConfigUser = true;
-      } catch (err) {
-        // do nothing here, it just means the user
-        // didn't define gatsby-config.js
-      }
-
-      // inject gatsby-config.js wrapper
-      await writeFilePromise(
-        join(entrypointDir, 'gatsby-config.js'),
-        createGatsbyConfig(gatsbyConfigUser),
-        { encoding: 'utf-8' }
-      );
-
-      const gatsbyPluginNowPath = join(
-        entrypointDir,
-        'plugins',
-        'gatsby-plugin-now'
-      );
-
-      await mkdirPromise(join(entrypointDir, 'plugins'));
-      await mkdirPromise(gatsbyPluginNowPath);
-
-      await copyFilePromise(
-        require.resolve('./gatsby-plugin-now/gatsby-node.js'),
-        join(gatsbyPluginNowPath, 'gatsby-node.js')
-      );
-      await copyFilePromise(
-        require.resolve('./gatsby-plugin-now/package.json'),
-        join(gatsbyPluginNowPath, 'package.json')
-      );
+      await injectGatsbyConfig(entrypointDir);
     },
   },
   {
