--- packages/now-static-build/src/gatsby-generators.ts
@@ -1,19 +1,0 @@
-export function createGatsbyConfig(gatsbyConfigUser: boolean) {
-  return `let userConfig = {}
-
-${
-  gatsbyConfigUser
-    ? `try {
-  userConfig = require('./gatsby-config-user.js')
-} catch (err) {}`
-    : ''
-}
-
-module.exports = {
-  ...userConfig,
-  plugins: [
-    ...(userConfig.plugins || []),
-    'gatsby-plugin-now'
-  ]
-}`;
-}
